+++ src/connection/adapter/http_adapter.ts
@@ -9,7 +9,6 @@ export class HttpAdapter extends BaseHttpAdapter implements Connection {
   constructor(config: ConnectionParams, logger: LogWriter) {
     const agent = new Http.Agent({
       keepAlive: true,
-      timeout: config.request_timeout,
       maxSockets: config.max_open_connections,
     })
     super(config, logger, agent)
@@ -17,12 +16,13 @@ export class HttpAdapter extends BaseHttpAdapter implements Connection {
 
   protected createClientRequest(
     url: URL,
-    params: RequestParams
+    params: RequestParams & { abort_controller: AbortController }
   ): Http.ClientRequest {
     return Http.request(params.url, {
       method: params.method,
       agent: this.agent,
       headers: this.getHeaders(params),
+      signal: params.abort_controller.signal,
     })
   }
 }