+++ src/schema/table.ts
@@ -36,13 +36,13 @@ export interface SelectOptions<S extends Shape> {
   where?: WhereExpr<S>
   order_by?: NonEmptyArray<[keyof S, 'ASC' | 'DESC']>
   clickhouse_settings?: ClickHouseSettings
-  abort_signal?: AbortSignal
+  abort_controller?: AbortController
 }
 
 export interface InsertOptions<S extends Shape> {
   values: Infer<S>[] | InsertStream<Infer<S>>
   clickhouse_settings?: ClickHouseSettings
-  abort_signal?: AbortSignal
+  abort_controller?: AbortController
 }
 
 export class Table<S extends Shape> {
@@ -57,18 +57,19 @@ export class Table<S extends Shape> {
     const { stream } = await this.client.exec({
       query,
       clickhouse_settings: options.clickhouse_settings,
+      returnResponseStream: true,
     })
     return stream
   }
 
   async insert({
-    abort_signal,
+    abort_controller,
     clickhouse_settings,
     values,
   }: InsertOptions<S>): Promise<void> {
     await this.client.insert({
       clickhouse_settings,
-      abort_signal,
+      abort_controller,
       table: getTableName(this.options),
       format: 'JSONEachRow',
       values,
@@ -76,7 +77,7 @@ export class Table<S extends Shape> {
   }
 
   async select({
-    abort_signal,
+    abort_controller,
     clickhouse_settings,
     columns,
     order_by,
@@ -86,7 +87,7 @@ export class Table<S extends Shape> {
     const rs = await this.client.query({
       query,
       clickhouse_settings,
-      abort_signal,
+      abort_controller,
       format: 'JSONEachRow',
     })